refactor: renames case to sample in cli
